+++ qgis-app/styles/templates/styles/style_base.html
@@ -70,9 +70,7 @@
             $("form.navbar-search").attr("action", "/styles/");
             })
     </script>
-
-<hr>
-<h1>Helloo</h1>
+    
 {% endblock %}
 
 {% block "credits" %}
